pyproject.toml
[tool.poetry]
name = "tffsmount"
description = "Read-only mounter for tffs filesystem"
version = "0.1"
readme = "README.md"
license = "Apache-2.0"
authors = ["Francis Hoogendijk <f.hoogendijk@nfi.nl>"]
repository = "https://github.com/NetherlandsForensicInstitute/tffsmount"

[tool.poetry.dependencies]
python = "^3.7"
fusepy = "^3.0.1"
kaitaistruct = "^0.10"
crcengine = "^0.3"
pytest = { version = "^7.2", optional = true }

[tool.poetry.extras]
test = ["pytest"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"